Add example for getting data from a single Wikidata item
This example shows how to get data from a single Wikidata item. This
is useful when you can map your pages to Wikidata items and want
to pull some information from Wikidata to fill the page. The advantage
is that the data is curated and maintained by Wikidata and will
hopefully be accurate and up to date.